requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.64.0

# File handling and text extraction
textract>=1.6.5
PyMuPDF>=1.23.0
python-docx>=0.8.11
openpyxl>=3.1.0
pdfplumber>=0.9.0

# NLP and embeddings
sentence-transformers>=2.2.2
transformers>=4.21.0
torch>=1.13.0
scikit-learn>=1.2.0
spacy>=3.4.0
nltk>=3.8.0

# Microsoft Graph API (optional)
msal>=1.20.0
requests>=2.28.0

# Web interface (optional)
streamlit>=1.25.0
plotly>=5.15.0
dash>=2.10.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
colorama>=0.4.6
rich>=13.0.0

# Development and testing
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.10.0
flake8>=5.0.0
mypy>=0.991

# SMB/Network file access
smbprotocol>=1.10.0
pysmb>=1.2.8

# Hashing and security
hashlib-compat>=1.0.1

